+++ packages/f-header/CHANGELOG.md
@@ -4,6 +4,17 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 
+v2.0.0-beta-21
+------------------------------
+*October 10, 2019*
+
+### Changed
+ - The way data is passed around in the order count logic
+
+### Added
+ - Tests for order count logic
+
+
 v2.0.0-beta-20
 ------------------------------
 *October 9, 2019*